Moved view caching to separate file
